修复时钟校准的数据类型

+++ soc/hdsc/drivers/include/clktrim/am_hc32_clktrim.h
@@ -47,10 +47,10 @@ extern "C" {
  * \return <0，被监控时钟失效
  *        AM_OK，工作正常
  */
-uint8_t am_hc32_clktrim_monitor (uint16_t                     rcntval,
-                                   uint16_t                     ccntval,
-                                   amhw_clktrim_refclk_sel_t    refclk_sel,
-                                   amhw_clktrim_calclk_sel_t    calclk_sel);
+uint8_t am_hc32_clktrim_monitor (uint32_t                     rcntval,
+                                 uint32_t                     ccntval,
+                                 amhw_clktrim_refclk_sel_t    refclk_sel,
+                                 amhw_clktrim_calclk_sel_t    calclk_sel);
 
 /**
  * \brief CLKTRIM校准模式配置
@@ -59,12 +59,11 @@ uint8_t am_hc32_clktrim_monitor (uint16_t                     rcntval,
  * \param[in] refclk_sel : 参考时钟选择
  * \param[in] calclk_sel : 待校准时钟选择
  *
- * \return <0，需调整校准时间
- *         >0，校准时钟计数器的值 / 参考时钟计数器的值 * 1000
+ * \return 校准时钟计数器的值
  */
-uint8_t am_hc32_clktrim_calibrate (uint16_t                     rcntval,
-                                     amhw_clktrim_refclk_sel_t    refclk_sel,
-                                     amhw_clktrim_calclk_sel_t    calclk_sel);
+uint32_t am_hc32_clktrim_calibrate (uint32_t                     rcntval,
+                                   amhw_clktrim_refclk_sel_t    refclk_sel,
+                                   amhw_clktrim_calclk_sel_t    calclk_sel);
 
 /**
  * @}
